refactor(Header.js) - persistencia de parametros seleccionado por usuaria
