+++ src/index.html
@@ -5,7 +5,7 @@
     <title>Dataverse Chat</title>
     <link rel="stylesheet" href="./styles.css">
     <link rel="stylesheet" href="./components/header/header.css">
-    <link rel="stylesheet" href="./pages/details.css">
+    <link rel="stylesheet" href="./pages/Details/details.css">
   </head>
   <body>
     <div id="root"></div>
