+++ src/pages/Details/details.css
